core-mwc: add missing explicit uint32_t cast

+++ core-mwc.c
@@ -160,7 +160,7 @@ void stress_mwc_reseed(void)
 		 *  Mix in some initial system values
 		 */
 		mwc.z ^= shim_rol32n(mwc.z, stress_get_cpu() & 0x1f);
-		mwc.w ^= shim_rol32n(mwc.w, stress_get_phys_mem_size() >> 22);
+		mwc.w ^= shim_rol32n(mwc.w, (uint32_t)(stress_get_phys_mem_size() >> 22));
 		mwc.z ^= stress_get_filesystem_size();
 		mwc.z ^= stress_get_kernel_release();
 		mwc.w ^= shim_rol32n((uint32_t)stress_get_ticks_per_second(), 3);
